move NamespaceStore to own module
- Namespace store is not generic over the Namespace anymore
- create method takes a config
- simplify loading namespaces
