+++ libsql-server/src/namespace/mod.rs
@@ -294,415 +294,6 @@ impl MakeNamespace for ReplicaNamespaceMaker {
 
 type NamespaceEntry<T> = Arc<RwLock<Option<Namespace<T>>>>;
 
-/// Stores and manage a set of namespaces.
-pub struct NamespaceStore<M: MakeNamespace> {
-    inner: Arc<NamespaceStoreInner<M>>,
-}
-
-impl<M: MakeNamespace> Clone for NamespaceStore<M> {
-    fn clone(&self) -> Self {
-        Self {
-            inner: self.inner.clone(),
-        }
-    }
-}
-
-struct NamespaceStoreInner<M: MakeNamespace> {
-    store: Cache<NamespaceName, NamespaceEntry<M::Database>>,
-    metadata: MetaStore,
-    /// The namespace factory, to create new namespaces.
-    make_namespace: M,
-    allow_lazy_creation: bool,
-    has_shutdown: AtomicBool,
-    snapshot_at_shutdown: bool,
-}
-
-impl<M: MakeNamespace> NamespaceStore<M> {
-    pub async fn new(
-        make_namespace: M,
-        allow_lazy_creation: bool,
-        snapshot_at_shutdown: bool,
-        max_active_namespaces: usize,
-        base_path: &Path,
-        meta_store_config: MetaStoreConfig,
-    ) -> crate::Result<Self> {
-        let metadata = MetaStore::new(meta_store_config, base_path).await?;
-        tracing::trace!("Max active namespaces: {max_active_namespaces}");
-        let store = Cache::<NamespaceName, NamespaceEntry<M::Database>>::builder()
-            .async_eviction_listener(move |name, ns, cause| {
-                tracing::debug!("evicting namespace `{name}` asynchronously: {cause:?}");
-                // TODO(sarna): not clear if we should snapshot-on-evict...
-                // On the one hand, better to do so, because we have no idea
-                // for how long we're evicting a namespace.
-                // On the other, if there's lots of cache pressure, snapshotting
-                // very often will kill the machine's I/O.
-                Box::pin(async move {
-                    tracing::info!("namespace `{name}` deallocated");
-                    // shutdown namespace
-                    if let Some(ns) = ns.write().await.take() {
-                        if let Err(e) = ns.shutdown(snapshot_at_shutdown).await {
-                            tracing::error!("error deallocating `{name}`: {e}")
-                        }
-                    }
-                })
-            })
-            .max_capacity(max_active_namespaces as u64)
-            .time_to_idle(Duration::from_secs(86400))
-            .build();
-
-        Ok(Self {
-            inner: Arc::new(NamespaceStoreInner {
-                store,
-                metadata,
-                make_namespace,
-                allow_lazy_creation,
-                has_shutdown: AtomicBool::new(false),
-                snapshot_at_shutdown,
-            }),
-        })
-    }
-
-    pub fn exists(&self, namespace: &NamespaceName) -> bool {
-        self.inner.metadata.exists(namespace)
-    }
-
-    pub async fn destroy(&self, namespace: NamespaceName) -> crate::Result<()> {
-        if self.inner.has_shutdown.load(Ordering::Relaxed) {
-            return Err(Error::NamespaceStoreShutdown);
-        }
-        let mut bottomless_db_id_init = NamespaceBottomlessDbIdInit::FetchFromConfig;
-        if let Some(ns) = self.inner.store.remove(&namespace).await {
-            // deallocate in-memory resources
-            if let Some(ns) = ns.write().await.take() {
-                bottomless_db_id_init = NamespaceBottomlessDbIdInit::Provided(
-                    NamespaceBottomlessDbId::from_config(&ns.db_config_store.get()),
-                );
-                ns.destroy().await?;
-            }
-        }
-
-        // destroy on-disk database and backups
-        self.inner
-            .make_namespace
-            .destroy(
-                namespace.clone(),
-                bottomless_db_id_init,
-                true,
-                &self.inner.metadata,
-            )
-            .await?;
-        tracing::info!("destroyed namespace: {namespace}");
-
-        Ok(())
-    }
-
-    pub async fn reset(
-        &self,
-        namespace: NamespaceName,
-        restore_option: RestoreOption,
-    ) -> anyhow::Result<()> {
-        // The process for reseting is as follow:
-        // - get a lock on the namespace entry, if the entry exists, then it's a lock on the entry,
-        // if it doesn't exist, insert an empty entry and take a lock on it
-        // - destroy the old namespace
-        // - create a new namespace and insert it in the held lock
-        let entry = self
-            .inner
-            .store
-            .get_with(namespace.clone(), async { Default::default() })
-            .await;
-        let mut lock = entry.write().await;
-        if let Some(ns) = lock.take() {
-            ns.destroy().await?;
-        }
-        // destroy on-disk database
-        self.inner
-            .make_namespace
-            .destroy(
-                namespace.clone(),
-                NamespaceBottomlessDbIdInit::FetchFromConfig,
-                false,
-                &self.inner.metadata,
-            )
-            .await?;
-        let ns = self
-            .inner
-            .make_namespace
-            .create(
-                namespace.clone(),
-                restore_option,
-                NamespaceBottomlessDbId::NotProvided,
-                self.make_reset_cb(),
-                &self.inner.metadata,
-            )
-            .await?;
-
-        lock.replace(ns);
-
-        Ok(())
-    }
-
-    // This is only called on replica
-    fn make_reset_cb(&self) -> ResetCb {
-        let this = self.clone();
-        Box::new(move |op| {
-            let this = this.clone();
-            tokio::spawn(async move {
-                match op {
-                    ResetOp::Reset(ns) => {
-                        tracing::info!("received reset signal for: {ns}");
-                        if let Err(e) = this.reset(ns.clone(), RestoreOption::Latest).await {
-                            tracing::error!("error resetting namespace `{ns}`: {e}");
-                        }
-                    }
-                    ResetOp::Destroy(ns) => {
-                        if let Err(e) = this.destroy(ns.clone()).await {
-                            tracing::error!("error destroying namesace `{ns}`: {e}",);
-                        }
-                    }
-                }
-            });
-        })
-    }
-
-    pub async fn fork(
-        &self,
-        from: NamespaceName,
-        to: NamespaceName,
-        timestamp: Option<NaiveDateTime>,
-    ) -> crate::Result<()> {
-        if self.inner.has_shutdown.load(Ordering::Relaxed) {
-            return Err(Error::NamespaceStoreShutdown);
-        }
-
-        // check that the source namespace exists
-        if !self.inner.metadata.exists(&from) {
-            return Err(crate::error::Error::NamespaceDoesntExist(from.to_string()));
-        }
-
-        let to_entry = self
-            .inner
-            .store
-            .get_with(to.clone(), async { Default::default() })
-            .await;
-        let mut to_lock = to_entry.write().await;
-        if to_lock.is_some() {
-            return Err(crate::error::Error::NamespaceAlreadyExist(to.to_string()));
-        }
-
-        let from_entry = self
-            .inner
-            .store
-            .try_get_with(from.clone(), async {
-                let ns = self
-                    .inner
-                    .make_namespace
-                    .create(
-                        from.clone(),
-                        RestoreOption::Latest,
-                        NamespaceBottomlessDbId::NotProvided,
-                        self.make_reset_cb(),
-                        &self.inner.metadata,
-                    )
-                    .await?;
-                tracing::info!("loaded namespace: `{to}`");
-                Ok::<_, crate::error::Error>(Arc::new(RwLock::new(Some(ns))))
-            })
-            .await?;
-
-        let from_lock = from_entry.read().await;
-        let Some(from_ns) = &*from_lock else {
-            return Err(crate::error::Error::NamespaceDoesntExist(to.to_string()));
-        };
-
-        let to_ns = self
-            .inner
-            .make_namespace
-            .fork(from_ns, to.clone(), timestamp, &self.inner.metadata)
-            .await?;
-
-        to_lock.replace(to_ns);
-
-        Ok(())
-    }
-
-    pub async fn with_authenticated<Fun, R>(
-        &self,
-        namespace: NamespaceName,
-        auth: Authenticated,
-        f: Fun,
-    ) -> crate::Result<R>
-    where
-        Fun: FnOnce(&Namespace<M::Database>) -> R + 'static,
-    {
-        if self.inner.has_shutdown.load(Ordering::Relaxed) {
-            return Err(Error::NamespaceStoreShutdown);
-        }
-        if !auth.is_namespace_authorized(&namespace) {
-            return Err(Error::NamespaceDoesntExist(namespace.to_string()));
-        }
-
-        self.with(namespace, f).await
-    }
-
-    pub async fn with<Fun, R>(&self, namespace: NamespaceName, f: Fun) -> crate::Result<R>
-    where
-        Fun: FnOnce(&Namespace<M::Database>) -> R + 'static,
-    {
-        let init = {
-            let namespace = namespace.clone();
-            async move {
-                if namespace != NamespaceName::default()
-                    && !self.inner.metadata.exists(&namespace)
-                    && !self.inner.allow_lazy_creation
-                {
-                    return Err(Error::NamespaceDoesntExist(namespace.to_string()));
-                }
-                let ns = self
-                    .inner
-                    .make_namespace
-                    .create(
-                        namespace.clone(),
-                        RestoreOption::Latest,
-                        NamespaceBottomlessDbId::NotProvided,
-                        self.make_reset_cb(),
-                        &self.inner.metadata,
-                    )
-                    .await?;
-                tracing::info!("loaded namespace: `{namespace}`");
-
-                Ok(Some(ns))
-            }
-        };
-
-        let f = {
-            let name = namespace.clone();
-            move |ns: NamespaceEntry<M::Database>| async move {
-                let lock = ns.read().await;
-                match &*lock {
-                    Some(ns) => Ok(f(ns)),
-                    // the namespace was taken out of the entry
-                    None => Err(Error::NamespaceDoesntExist(name.to_string())),
-                }
-            }
-        };
-
-        self.with_lock_or_init(namespace, f, init).await?
-    }
-
-    async fn with_lock_or_init<Fun, R, Init, Fut>(
-        &self,
-        namespace: NamespaceName,
-        f: Fun,
-        init: Init,
-    ) -> crate::Result<R>
-    where
-        Fun: FnOnce(NamespaceEntry<M::Database>) -> Fut,
-        Fut: Future<Output = R>,
-        Init: Future<Output = crate::Result<Option<Namespace<M::Database>>>>,
-    {
-        let before_load = Instant::now();
-        let ns = self
-            .inner
-            .store
-            .try_get_with(
-                namespace.clone(),
-                init.map_ok(|ns| Arc::new(RwLock::new(ns))),
-            )
-            .await?;
-        NAMESPACE_LOAD_LATENCY.record(before_load.elapsed());
-        Ok(f(ns).await)
-    }
-
-    pub async fn create(
-        &self,
-        namespace: NamespaceName,
-        restore_option: RestoreOption,
-        bottomless_db_id: NamespaceBottomlessDbId,
-    ) -> crate::Result<()> {
-        // With namespaces disabled, the default namespace can be auto-created,
-        // otherwise it's an error.
-        if self.inner.allow_lazy_creation || namespace == NamespaceName::default() {
-            tracing::trace!("auto-creating the namespace");
-        } else if self.inner.metadata.exists(&namespace) {
-            return Err(Error::NamespaceAlreadyExist(namespace.to_string()));
-        }
-
-        let name = namespace.clone();
-        let bottomless_db_id_for_init = bottomless_db_id.clone();
-        let init = async {
-            let ns = self
-                .inner
-                .make_namespace
-                .create(
-                    name.clone(),
-                    restore_option,
-                    bottomless_db_id_for_init,
-                    self.make_reset_cb(),
-                    &self.inner.metadata,
-                )
-                .await;
-            match ns {
-                Ok(ns) => {
-                    tracing::info!("loaded namespace: `{name}`");
-                    Ok(Some(ns))
-                }
-                // return an empty slot to put the new namespace in
-                Err(Error::NamespaceDoesntExist(_)) => Ok(None),
-                Err(e) => Err(e),
-            }
-        };
-
-        self.with_lock_or_init(namespace, |_| async { Ok(()) }, init)
-            .await?
-    }
-
-    pub async fn shutdown(self) -> crate::Result<()> {
-        self.inner.has_shutdown.store(true, Ordering::Relaxed);
-        for (_name, entry) in self.inner.store.iter() {
-            let mut lock = entry.write().await;
-            if let Some(ns) = lock.take() {
-                ns.shutdown(self.inner.snapshot_at_shutdown).await?;
-            }
-        }
-        self.inner.metadata.shutdown().await?;
-        self.inner.store.invalidate_all();
-        self.inner.store.run_pending_tasks().await;
-        Ok(())
-    }
-
-    pub(crate) async fn stats(&self, namespace: NamespaceName) -> crate::Result<Arc<Stats>> {
-        self.with(namespace, |ns| ns.stats.clone()).await
-    }
-
-    pub(crate) async fn config_store(
-        &self,
-        namespace: NamespaceName,
-    ) -> crate::Result<MetaStoreHandle> {
-        self.with(namespace, |ns| ns.db_config_store.clone()).await
-    }
-
-    pub(crate) fn meta_store(&self) -> MetaStore {
-        self.inner.metadata.clone()
-    }
-}
-
-/// A namespace isolates the resources pertaining to a database of type T
-#[derive(Debug)]
-pub struct Namespace<T> {
-    pub db: T,
-    name: NamespaceName,
-    /// The set of tasks associated with this namespace
-    tasks: JoinSet<anyhow::Result<()>>,
-    stats: Arc<Stats>,
-    db_config_store: MetaStoreHandle,
-}
-
-impl<T: Database> Namespace<T> {
-    pub(crate) fn name(&self) -> &NamespaceName {
-        &self.name
-    }
-
     async fn destroy(mut self) -> anyhow::Result<()> {
         self.tasks.shutdown().await;
         self.db.destroy();
